refactor(scanner): moved redux actions from interactionScreen to scanner container
